+++ src/router/router.tsx
@@ -1,9 +1,11 @@
 import { RouteObject } from 'react-router-dom'
 
 import MainLayout from '@/components/MainLayout'
+import FriendTimeTablePage from '@/pages/FriendTimeTablePage'
 import LandingPage from '@/pages/LandingPage'
 import Login from '@/pages/Login'
 import MyPage from '@/pages/MyPage'
+import MyTimeTablePage from '@/pages/MyTimeTablePage'
 import TimeTablePage from '@/pages/TimeTablePage'
 import ProtectedRoutes from '@/router/ProtectedRoutes'
 
@@ -17,7 +19,14 @@ const routes: RouteObject[] = [
         element: <ProtectedRoutes />,
         children: [
           { path: '', element: <MyPage /> },
-          { path: 'timetable', element: <TimeTablePage /> },
+          {
+            path: 'timetable',
+            element: <TimeTablePage />,
+            children: [
+              { path: '', element: <MyTimeTablePage /> },
+              { path: 'friend', element: <FriendTimeTablePage /> },
+            ],
+          },
         ],
       },
       { path: 'home', element: <LandingPage /> },
